# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PingCortex API
  description: API service for ping cortex app - it covers all endpoints and data structures
  version: 1.0.0
servers:
  - url: 'http://localhost:8080/api'
    description: Local development
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Create a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive access + refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange a refresh token for a new access token (rotates the refresh token)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        '401':
          description: Refresh token invalid, expired, or revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke the given refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        '204':
          description: Logged out
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: INVALID_CREDENTIALS
        message:
          type: string
          example: Email or password is incorrect.
      required: [code, message]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          example: alice@gmail.com
        password:
          type: string
          example: password123
      required: [email, password]
    JWTPayload:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        email:
          type: string
          example: alice@gmail.com
        level:
          type: string
          example: First year in informatics
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: edab48bf5...
        refreshToken:
          type: string
          example: edab48bf6...
        user:
          $ref: "#/components/schemas/JWTPayload"
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          example: edab48bf6...
      required: [refreshToken]
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
          example: edab48bf7...
        refreshToken:
          type: string
          description: New refresh token; the previous one is revoked (rotation).
          example: edab48bf8...
    UserRegister:
      type: object
      properties:
        email:
          type: string
          example: alice@gmail.com
        name:
          type: string
          example: Alice Wayne
        password:
          type: string
          example: password123
        level:
          type: string
          example: First year in informatics
      required: [email, name, password, level]
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426655440000
        email:
          type: string
          example: alice@gmail.com
        name:
          type: string
          example: Alice Wayne
        level:
          type: string
          example: First year in informatics
        alertThreshold:
          type: integer
          nullable: true
          example: 9000
        createdAt:
          type: string
          format: date-time
          example: 2026-09-04T15:27:00
...